Metrics (#4)
Replace previous method of pulling meetings and participants with data based approach using dashboards endpoint so each run pulls in a single day based on the last date data was pulled. To catch up the database this can mean it's necessary to run the job repeatedly until all prior dates have been loaded. After that it can be run on a daily basis.